refactor(backend): group catalog routes
