+++ apps/backend/src/catalog/distribution/public/media.test.ts
@@ -2,7 +2,7 @@ import assert from "node:assert/strict";
 import test from "node:test";
 import type pg from "pg";
 import type { DatabaseExecutor, SqlValue } from "../../../database";
-import { createCatalogPublicRoutes } from "../../../routes/catalogPublic";
+import { createCatalogPublicRoutes } from "../../../routes/catalog/public";
 import { HttpError } from "../../../shared/errors";
 import {
   loadPublicCatalogPackageMediaForDownloadInExecutor,
